pyproject.toml
[project]
name = "HOPE-vs-vit-research"
version = "0.1.0"
description = "Gold standard research repository comparing HOPE and ViT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.2.0",
    "lightning>=2.2.0",
    "hydra-core>=1.3.2",
    "hydra-colorlog>=1.2.0",
    "hydra-optuna-sweeper>=1.2.0",
    "wandb>=0.16.0",
    "python-dotenv>=1.0.0",
    "pyrootutils>=1.0.4",
    "rich>=13.7.0",
    "numpy<2.0.0", # Pin to avoid 2.0 conflicts with some DL libraries
    "timm>=0.9.0", # For drop_path and other vision utilities
    "psutil",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pre-commit>=3.6.0",
    "ruff>=0.3.0",  # uv works great with ruff
    "ipykernel",
    "matplotlib",
    "scikit-learn",
    "vastai"
]
# Extra dependencies for standard comparison architectures (Transformer++, RetNet, GatedDeltaNet).
# Install with: uv sync --extra comparison-architectures
comparison-architectures = [
    "einops",
    "flash-attn>=2.0.0",
    "causal-conv1d>=1.2.0",
    "flash-linear-attention",  # fla — required by GatedDeltaNet
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
